+++ ecdklint/aws_constants/regions.py
@@ -13,4 +13,40 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-REGIONS = {"eu-central-1": "euc1", "eu-west-1": "euw1", "us-east-1": "use1"}
+# This should list all available AWS regions.
+REGIONS = {
+    "af-south-1": "afs1",
+    "ap-east-1": "ape1",
+    "ap-northeast-1": "apn1",
+    "ap-northeast-2": "apn2",
+    "ap-northeast-3": "apn3",
+    "ap-south-1": "aps1",
+    "ap-south-2": "aps2",
+    "ap-southeast-1": "apo1",
+    "ap-southeast-2": "apo2",
+    "ap-southeast-3": "apo3",
+    "ap-southeast-4": "apo4",
+    "ap-southeast-5": "apo5",
+    "ap-southeast-7": "apo7",
+    "ca-central-1": "cac1",
+    "ca-west-1": "caw1",
+    "eu-central-1": "euc1",
+    "eu-central-2": "euc2",
+    "eu-north-1": "eun1",
+    "eu-south-1": "eus1",
+    "eu-south-2": "eus2",
+    "eu-west-1": "euw1",
+    "eu-west-2": "euw2",
+    "eu-west-3": "euw3",
+    "il-central-1": "ilc1",
+    "me-central-1": "mec1",
+    "me-south-1": "mes1",
+    "mx-central-1": "mxc1",
+    "sa-east-1": "sae1",
+    "us-east-1": "use1",
+    "us-east-2": "use2",
+    "us-gov-east-1": "uge1",
+    "us-gov-west-1": "ugw1",
+    "us-west-1": "usw1",
+    "us-west-2": "usw2",
+}