+++ ecdklint/rules/__init__.py
@@ -14,9 +14,13 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from ecdklint.rules import (dummy, parameters_missing, region_invalid,
-                            region_missing, region_stackname_mismatch,
-                            stack_module_missing, stackname_too_long)
+from ecdklint.rules import (dummy, json5_file_extension_deprecated,
+                            parameters_missing, region_invalid,
+                            region_mismatch_stack_tags, region_missing,
+                            region_stackname_mismatch, stack_class_missing,
+                            stack_module_missing,
+                            stackname_doesnt_match_module_and_class,
+                            stackname_too_long)
 
 _RULES = {
     region_missing.ID: region_missing,
@@ -26,6 +30,10 @@ _RULES = {
     dummy.ID: dummy,
     parameters_missing.ID: parameters_missing,
     stackname_too_long.ID: stackname_too_long,
+    stack_class_missing.ID: stack_class_missing,
+    stackname_doesnt_match_module_and_class.ID: stackname_doesnt_match_module_and_class,
+    json5_file_extension_deprecated.ID: json5_file_extension_deprecated,
+    region_mismatch_stack_tags.ID: region_mismatch_stack_tags,
 }